requirements.txt
fastapi
uvicorn
SQLAlchemy
PyJWT
pytest
alembic
pydantic
python-dotenv
bcrypt==4.0.1
